requirements.txt
pandas
newsapi-python
python-dotenv
matplotlib
seaborn